fix return type of calculate score. for scoreOfWordInBetween, consider a new use case where a top word is also a top word in the returned structure (same for buttom word), fix infinite loop caused by that

+++ src/main/java/com/sv/score/dto/WordScoreDto.java
@@ -7,12 +7,12 @@ public class WordScoreDto implements Serializable {
     private static final long serialVersionUID = 4073797816654019357L;
 
     private String keyWord;
-    private Integer score;
+    private float score;
 
     public WordScoreDto() {
     }
 
-    public WordScoreDto(String keyWord, Integer score) {
+    public WordScoreDto(String keyWord, float score) {
         this.keyWord = keyWord;
         this.score = score;
     }
@@ -29,11 +29,11 @@ public class WordScoreDto implements Serializable {
         this.keyWord = keyWord;
     }
 
-    public Integer getScore() {
+    public float getScore() {
         return score;
     }
 
-    public void setScore(Integer score) {
+    public void setScore(Long score) {
         this.score = score;
     }
 }
